+++ src/utils.h
@@ -1,6 +1,8 @@
 #ifndef UTILS_H
 #define UTILS_H
 
+#include <stdint.h>
+
 #include <sys/time.h>
 
 #define US_PER_SECOND 1000000
@@ -11,13 +13,13 @@
  * @param tv time
  * @return time in us
  */
-unsigned long long int time_to_us(struct timeval tv);
+uint64_t time_to_us(struct timeval tv);
 
 /**
  * get time in us
  *
  * @return time in us
  */
-unsigned long long int get_time_us();
+uint64_t get_time_us();
 
 #endif
